requirements.txt
openai
streamlit
llama-index-core
llama-index-vector_stores-chroma
llama-index-embeddings-dashscope
chromadb
pdf2image